# lib/api-zod/src/generated/types/collaborator.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface Collaborator {
  id: number;
  userId: number;
  name: string;
  country: string;
  code: string;
  commissionPercent: string;
  active: boolean;
  createdAt: Date;
}