# lib/api-zod/src/generated/types/collaboratorDashboard.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { CollaboratorMonth } from "./collaboratorMonth";
import type { MoneyByCurrency } from "./moneyByCurrency";

export interface CollaboratorDashboard {
  name: string;
  code: string;
  link: string;
  commissionPercent: string;
  referrals: number;
  activeSubscriptions: number;
  newSubscriptionsThisMonth: number;
  cancellationsThisMonth: number;
  generatedThisMonth: MoneyByCurrency;
  pending: MoneyByCurrency;
  paid: MoneyByCurrency;
  history: CollaboratorMonth[];
}