# lib/api-zod/src/generated/types/collaboratorInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface CollaboratorInput {
  name: string;
  email: string;
  /** @minLength 8 */
  password: string;
  country: string;
  code: string;
  commissionPercent: string;
}