# lib/api-zod/src/generated/types/collaboratorMonth.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { MoneyByCurrency } from "./moneyByCurrency";

export interface CollaboratorMonth {
  month: string;
  referrals: number;
  newSubscriptions: number;
  cancellations: number;
  generated: MoneyByCurrency;
  pending: MoneyByCurrency;
  paid: MoneyByCurrency;
}