# lib/api-zod/src/generated/types/referralValidity.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface ReferralValidity {
  valid: boolean;
  code: string;
}