# lib/api-zod/src/generated/types/saasReceipt.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface SaasReceipt {
  id: number;
  professionalUserId: number;
  amount: string;
  currency: string;
  periodFrom: Date;
  periodTo: Date;
  receivedAt: Date;
  /** @nullable */
  reference: string | null;
  idempotencyKey: string;
  createdByUserId: number;
  /** @nullable */
  collaboratorId: number | null;
  /** @nullable */
  commissionPercentSnapshot: string | null;
  /** @nullable */
  commissionAmount: string | null;
  /** @nullable */
  paidAt: Date | null;
  /** @nullable */
  paidByUserId: number | null;
  /** @nullable */
  paymentReference: string | null;
  createdAt: Date;
}