# lib/api-zod/src/generated/types/saasReceiptInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface SaasReceiptInput {
  professionalUserId: number;
  amount: string;
  currency: string;
  periodFrom: Date;
  periodTo: Date;
  receivedAt: Date;
  reference?: string;
  idempotencyKey: string;
}