# lib/api-zod/src/generated/types/saasStatusInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { SaasStatusInputStatus } from "./saasStatusInputStatus";

export interface SaasStatusInput {
  professionalUserId: number;
  status: SaasStatusInputStatus;
  effectiveDate?: Date;
}