# lib/api-zod/src/generated/types/saasStatusResult.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Neurometric Lab API
 * OpenAPI spec version: 0.1.0
 */
import type { SaasStatusResultEvent } from "./saasStatusResultEvent";

export interface SaasStatusResult {
  professionalUserId: number;
  status: string;
  /** @nullable */
  event: SaasStatusResultEvent;
}